requirements.txt
Flask>=2.0
gunicorn
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
python-dotenv
pypdf
docx2txt
